requirements.txt
pybullet
torch
PyOpenGL
glfw
PyGLM
numpy
pydantic

# jupyter
jupyter

# pre-commit
pre-commit
pytest
mypy
